Update patch set 5
Patch Set 5: Code-Review-1

Couple of questions/issues:
1. Why are headers just added to files? I see a few 2 line changes without associated changes. If this is fixing a warning, please identify it as such and prehaps put it in a commit that shows that. Otherwise, it looks like the compile is going too fast.

2. It looks like the Android startup files are imported from somewhere. Can you please break it up into 2 commits:
a) A pure import as-is from whatever upstream. Indicate in the commit where it came from (ASOP/machine/manifest used/git hash/etc). This is needed so future upstream changes can be merged easily.
b) A second commit showing exactly what the local changes are. This is so the change doesn't get lost in a jungle of adds. Right now, it is unclear what the changes are and how it differs from upstream.

Patch-set: 5
Label: Code-Review=-1