Update patch set 3
Patch Set 3: Code-Review-1

Arrgg... As I have said in prior mails, this version of the driver is outdated. Please use the latest version with the components factored out. Support for this should be added to the OSPLib + output layer. The output layer would create the sysfs attributes. OSPLib is responsible for encoding it as a packet.

Patch-set: 3
Label: Code-Review=-1