Update patch set 1
Patch Set 1:

There are too many design change requests that are part of review comments. If that's the case, we need a comprehensive review of the original design. We can go on ranting about everything that is produced in a particular patch. 

However to resolve this, Hanyue, since you've originated this design please produce the following -
1. Requirement specification for OSP Daemon (includes functional, platform, device requirements)
2. Design spec that describes so called Driver model.

This can be referred by open source developers or customers when they'd like to do some ammends.

Specifically, make sure your design also mentions -
1. How this driver model is scalable if input or output driver IPC is not based on FS
2. How this driver model factors in algorithms (*this is close to the requirement mentioned in point 1)

Patch-set: 1