Update patch set 1
Patch Set 1:

It is a bit late to be asking for such things during a review. If you insist on written documentation, please open a proper issue request in the github repo and assign it to me.

There is nothing in the design that forces drivers to use anything specific to the filesystem. For what you are trying to accomplish - similar functionality to the daemon but in the runtime context of the HAL, it can be accomplished without breaking anything or introducing all these problems. As an initial pass (not an optimal solution):
1. Create an alternative to main as the thread starting point.
2. Create an output type driver to speak to talk to the HAL.

To address your questions:
1. There is nothing in the design that requires anything to do at all with the filesystem. If it is not clear, despite the name "file descriptor", it is not restricted to a file system. For example, this model works fine if you decide to use network services. The IIO driver is an example of a "slow" file descriptor; the CVS stuff is an example of a "fast" file descriptor. 
The alg's used in the Daemon should be API identical to the version used on the hub.

2. The driver model has nothing to do with the Algs. Algs should be API compatible to the same one as running on the hub.

Patch-set: 1