Update patch set 4
Patch Set 4:

A simple of implementing the enable/disable is to refactor OSPDaemon_pm.c. I suggest something like this:
Commit 1:
Seperate out the fd support. This should be a very simple change. Put the fd specific code a specific function. It should probally do something like on call back, convert name to sensor type. Create a global entry point that takes sensor type and acts on it. It should do almost the same thing as the current static function except it does the sensor type lookup. Something like
int OSPDaemon_power_control(ASensorType_t sensor, int state)

It will incur an added lookup for the daemon case but this is not something that should not happen frequently.

This should be a functionally equivalent change and can be verified with tests.

Commit 2 - Call the sensor type entry point from the HAL.

This way the code can be fully tested after commit 1.

Patch-set: 4