Update patch set 7
Patch Set 7: Code-Review-1

This is completely unacceptable. There has been too much trash code that demonstrate a lack of understanding and proper testinging in any of the patches. See comments in the other patches for specifics. Please update to the correct code base. You have had over a week (Since Dec 8 documented here and slighly earlier through other medium) to correct this. Continuing on this is a total waste of time. 

As the original author of this driver, I am putting it on the record this is inappropriate to continue on this code base and expect it to be merged.

Patch-set: 7
Reviewer: hy-github@hy-research.com <1006955@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1