Update patch set 7
Patch Set 7:

My earlier comment was FYI. This is discussed with Rian. Batching is one of most sought after deliverables, and it can't be blocked just because one individual does not accept it.
BTW,  Please explain what do you mean by this - "code that demonstrate a lack of understanding and proper testinging in any of the patches"?

This code has been tested for batching, enable with CTS. Also don't mix from other patches to this one.

Also you are not supposed to comment on what's being done from Dec 8. I think it's highly inappropriate. However, fyi, this patchset has been around from Nov-26, so you'd ample of time to review and give suggestions.

We've been busy integrating batching firmware and that's done in not more than a week's time. Merging of this patchset is accepted by Rian. Also, the code in feature-rpmsg branch is not reviewed and merged. That's not a normal practice.
If you want to get your source code reviewed and merged please resubmit your patch in this repository, or allow other developers to port feature-rpmsg to refactored repository.

Patch-set: 7