Implementation of Queue Driver
+ This is addresses one of the major review comments obatined for the
earlier commit
+ The queue driver is a dummy driver type which does not send the data
anywhere. When the send is called, it checks whether it has data
available in the queue and then posts on the semaphore.
+ The other change is in the OSPDaemon_get_sensor_data, after one sample
has been pulled from the queue, we check if we have more data in the
queue, if we do then we post on the semaphore, if we don't do this then
the sensor hal will not call get sensor sample until the next read from
the input driver - this works when the sensors are running in continuous
mode but it fails to fetch multiple packets that were queued in the
single read from the driver.

JIRA ID : MOT-2468

Change-Id: I21e19f62a924268a31865631e482ce9ec26772ce
Signed-off-by: Ravindra GV <rvijayakumar@audience.com>