Imporved poll call to fill in more sensor data
+ In the poll, only one data packet per sensor was filled into the array
sent by Android. In case we have more data packets for a particular we
will have to wait for the next poll call to send that data. This ensures
that we always try to fill as much sensor data packets as possible of
each sensor in one poll call. NOTE - The sensor data is still
interleaved this will result in multiple checks for a particular sensor
to check whether it has data.
+ Added more debug logs to help analyze the CTS test cases.

Change-Id: I73d2da11c5b5bcc14d4ae46a6ee7a6e46f83d880
Signed-off-by: Ravindra GV <rvijayakumar@audience.com>